remove config.validate() explicit calls, use post_init everywhere, all configs own file
