+++ src/dr_plotter/configs/__init__.py
@@ -1,10 +1,12 @@
-from dr_plotter.configs.legend_config import LegendConfig, LegendStrategy
-from dr_plotter.configs.figure_config import FigureConfig
 from dr_plotter.configs.cycle_config import CycleConfig
-from dr_plotter.configs.grouping_config import GroupingConfig
-from dr_plotter.configs.positioning_config import PositioningConfig
 from dr_plotter.configs.faceting_config import FacetingConfig
-from dr_plotter.configs.plot_config import LayoutConfig, StyleConfig, PlotConfig
+from dr_plotter.configs.figure_config import FigureConfig
+from dr_plotter.configs.grouping_config import GroupingConfig
+from dr_plotter.configs.layout_config import LayoutConfig
+from dr_plotter.configs.legend_config import LegendConfig, LegendStrategy
+from dr_plotter.configs.plot_config import PlotConfig
+from dr_plotter.configs.positioning_config import PositioningConfig
+from dr_plotter.configs.style_config import StyleConfig
 
 __all__ = [
     "CycleConfig",
